requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
python-dotenv
Flask>=2.0.0
asgiref>=3.0.0
